fix : A bug which is occurred when net name has unrecognized symbols in input CSV file process has been fixed.
D:\1_Work\20220106_DDR_Compliance\0_DB\Galileo_NetClassification_Fail_2133.csv
Take one of the net names of the above input files as an example,
V(M_DQ_6__D2_G83568-001_U1B5) [V],
the "-" symbol exists in the net name, which the AEDT script does not recognize.
Updated to replace the "-" symbol with the "_" using the replace function.

[Additional Update List]
1. Compliance GUI Update - not yet linked to the previous process
2. Concecutive Analysis Bug Fixed - For CSV input file Process, the error fixed that occurred after non-group analysis and during group analysis